requirements.txt
art
babel
charset-normalizer
config
docopt
flask-markdown
frozen-flask
gensim
gevent
humanize
matplotlib
nltk
pandas
pip-chill
python-version
sacremoses
sklearn
soundfile
tensorflow-estimator
tensorflow
torchaudio
torchtext
torchvision
transformers
